requirements.txt
pydglab-ws>=0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
websockets>=12.0
qrcode>=7.4.2
Pillow>=10.0.0
netifaces>=0.11.0
